% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_plots.R
\name{addalpha}
\alias{addalpha}
\title{Add alpha}
\usage{
addalpha(colors, alpha = 1)
}
\arguments{
\item{colors}{list of colors}

\item{alpha}{value of alpha}
}
\description{
Internal to colorRampPaletteALpha but exported in case a user wants to modify
}
