% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{cluster.covex}
\alias{cluster.covex}
\title{Computing a clustered covariance explained matrix}
\usage{
cluster.covex(object, clust.method = "ward.D", dist.method = "manhattan",
  plot = FALSE, ...)
}
\arguments{
\item{object}{mvtb output}

\item{clust.method}{clustering method for rows and columns. See \code{?hclust}}

\item{dist.method}{method for computing the distance between two lower triangluar covariance matrices. See \code{?dist} for alternatives.}

\item{plot}{Produces a heatmap of the covariance explained matrix. see \code{?heat.covex}}

\item{...}{Arguments passed to \code{heat.covex}}
}
\value{
clustered covariance matrix, with rows and columns.
}
\description{
For each pair of predictors, computes the distance between the correlation matrices of the outcomes explained by those predictors.
Values smaller than \code{getOption("digits")} are truncated to 0.
}
\seealso{
\code{heat.covex}
}

