% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_plots.R
\name{colorRampPaletteAlpha}
\alias{colorRampPaletteAlpha}
\title{Add alpha and ramps between colors}
\usage{
colorRampPaletteAlpha(colors, n = 32, interpolate = "linear")
}
\arguments{
\item{n}{number of colors}

\item{interpoloate}{linear, otherwise spline interpolation is used}
}
\description{
Internal to mvtb.heat, but exported for easy modification
}

