% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_cverr.R
\name{gbm.cverr}
\alias{gbm.cverr}
\title{Tune gbm via cross-validation to find the best set of metaparameters}
\usage{
gbm.cverr(x, y, distribution = "gaussian", cv.folds = 5, nt.start = 1000,
  nt.inc = 1000, verbose = T, w = NULL, var.monotone = NULL,
  interaction.depth = 1, n.minobsinnode = 10, shrinkage = 0.001,
  bag.fraction = 0.5, n.cores = 1, max.time = NULL, seed = NULL)
}
\arguments{
\item{x}{A \eqn{n x p} matrix or data frame of predictors.}

\item{y}{A \eqn{n x 1} matrix or vector corresponding to the observed 
outcome.}

\item{distribution}{The distribution to use when fitting each \code{gbm}
model. For continuous outcomes, the available distributions are "gaussian" 
(squared error) and "laplace" (absolute loss). For dichotomous outcomes, the
available distributions are  "bernoulli" (logistic regression for 0-1 
outcomes) and "adaboost" (the AdaBoost exponential loss for 0-1 outcomes). 
Finally, the "poisson" distribution is available for count outcomes.}

\item{cv.folds}{Number of cross-validation folds to perform.}

\item{nt.start}{Initial number of trees used to model y.}

\item{nt.inc}{Number of trees incrementally added until the cross-validation
error is minimized.}

\item{verbose}{If TRUE, then \code{gbm.cverr} will print status information
to the console.}

\item{w}{a vector of weights of the same length as y. NOTE: to
evaluate the effect of different weight vectors, a list can be
passed to w in which each element follows the structure described
above.}

\item{var.monotone}{an optional vector, the same length as the number of 
predictors, indicating which variables have a monotone increasing (+1), 
decreasing (-1), or arbitrary (0) relationship with the outcome. NOTE: to
evaluate the effect of different monotonicity constraints, a list can be
passed to var.monotone in which each element follows the structure described
above.}

\item{interaction.depth}{The maximum depth of variable interactions: 1 builds
an additive model, 2 builds a model with up to two-way interactions, etc. 
NOTE: Multiple values can be passed in a vector to evaluate the 
cross-validation  error using multiple interaction depths.}

\item{n.minobsinnode}{The minimum number of observations (not total weights) 
in the terminal nodes of the trees. NOTE: Multiple values can be passed in a
vector to evaluate the cross-validation error using multiple minimum node 
sizes.}

\item{shrinkage}{A shrinkage parameter applied to each tree in the expansion.
Also known as the learning rate or step-size reduction. NOTE: Multiple values
can be passed in a vector to evaluate the cross-validation error using 
multiple shrinkage penalties.}

\item{bag.fraction}{The fraction of independent training observations (or 
patients) randomly selected to propose the next tree in the expansion, 
depending on the obs.id vector multiple training data rows may belong to a 
single 'patient'. This introduces randomness into the model fit. NOTE: 
Multiple values can be passed in a vector to evaluate the cross-validation 
error using multiple bag fractions.}

\item{n.cores}{Number of cores that will be used to estimate cross-validation
folds in parallel. Only available on linux-based machines.}

\item{max.time}{Maximum number of seconds that the model will continue adding
trees for a given set of metaparameters. This optional argument allows users 
to find the best possible solution in scenarios characterized by limited 
computational resources.}

\item{seed}{Seed that will guarantee \code{gbm.cverr} to produce identical 
results across multiple runs. Utilizing \code{set.seed} prior to calling 
\code{gbm.cverr} does NOT ensure equal results if \code{bag.fraction < 1}}
}
\value{
An object with XXXXX elements and a summary function. Calling
\code{summary(gbm.cverr.res)} produces a data frame comprised of:
\item{X}{fill this in}
\item{XX}{fill this in}
\item{XXX}{fill this in}
In addition to the information in the three columns comprising
\code{summary(gbm.cverr.res)}, the \code{gbm.cverr.res} object also contains:

\item{w}{A list of the weights passed to \code{gbm.cverr} corresponding to 
each set of metaparameters listed in the table of results. This will only be 
included if a list was passed to \code{w} when calling \code{gbm.cverr}.}
\item{var.monotone}{A list of the monotonicity constraints passed to 
\code{gbm.cverr} corresponding to each set of metaparameters listed in the 
table of results. This will only be included if a list was passed to 
\code{var.monotone} when calling \code{gbm.cverr}.}
\item{cv.err}{A list of the cross-validation error across all trees 
corresponding to each set of metaparameters listed in the table of results.}
\item{res}{The raw data frame upon which \code{summary(gbm.cverr.res)} is 
based.}
}
\description{
This function serves two purposes. First, it computes the cross-validation
error across a grid of \code{gbm} metaparameters input by the user, allowing
the model to be easily tuned for a given problem. This process can be 
executed in parallel on linux-based machines. Second, it alleviates the 
burden of selecting a maximum number of trees for a given set of 
metaparameters by allowing the algorithm to run until the best
number of trees has been selected according to the cross-validation error, in
contrast to the standard approach to \code{gbm}, in which a maximum
\code{n.trees} must also be tuned. This allows users to avoid two
types of problem associated with an inappropriate selection for
\code{n.trees}: (1) failing to specify enough trees and therefore using a 
sub-optimal model, and (2) specifying far more trees than are necessary, 
therefore making the \code{gbm} run for far more time than necessary.
}
\details{
The main output of \code{gbm.cverr} is a data frame with rows corresponding
to sets of metaparameters and columns corresponding to (1) the values 
defining each set of metaparameters, (2) the minimum cross-validation error
corresponding to each row, and (3) the number of trees that yielded the 
minimum cross-validation error in each row. These results are intended to 
allow users to make informed decisions about the metaparameters passed to 
\code{gbm} when fitting the model that will be interpreted and/or used for
prediction in the future.

Note that the metaparamter values passed to \code{w}, \code{var.monotone}, 
\code{interaction.depth}, \code{n.minobsinnode}, \code{shrinkage}, and 
\code{bag.fraction} will be fully crossed and evaluated.
}
\examples{
data(wellbeing)
y <- wellbeing[,25]
x <- wellbeing[,1:20]

mm <- gbm.cverr(x = x, y = y, 
               distribution = 'gaussian', 
               cv.folds = 2, 
               
               nt.start = 100, 
               nt.inc = 100, 
               max.time = 1, 
               
               seed = 12345,
               interaction.depth = c(1, 5), 
               shrinkage = 0.01,
               n.minobsinnode = c(5, 50), 
               verbose = TRUE)

summary(mm)

}
\author{
Daniel B. McArtor (dmcartor@nd.edu)
}

