% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{gbm.ri}
\alias{gbm.ri}
\title{Computes the relative influence of each predictor for each outcome from gbm.}
\usage{
gbm.ri(out, n.trees = NULL, relative = "col", ...)
}
\arguments{
\item{out}{mvtb output object}

\item{n.trees}{number of trees to use}

\item{relative}{If 'col', each column sums to 100. If 'tot', the whole matrix sums to 100 (a percent). If 'n', the raw reductions in SSE are returned.}

\item{...}{Additional arguments passed to \code{gbm::relative.influence}}
}
\value{
Matrix of (relative) influences.
}
\description{
The relative influence of a predictor is the reduction in sums of squares attributable to splits on individual predictors.
It is often expressed as a percent (sums to 100).
}

