% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_grid.R
\name{gbm_grid}
\alias{gbm_grid}
\title{CV Tune gbm fits over a grid of parameters in parallel}
\usage{
gbm_grid(y, x, cv.folds, mc.cores = 1, subset = NULL, ...)
}
\arguments{
\item{y}{vector of outcomes}

\item{x}{matrix of predictors}

\item{cv.folds}{number of cross-validation folds}

\item{mc.cores}{number of cores}

\item{subset}{subset of the data to use for training; cross-validation will be performed within subset}

\item{...}{args to gbm.fit. Arguments can be passed vectors. The function tunes across all rows of \code{expand.grid(...)}}
}
\value{
A list with \code{$gbm} (fitted model with best meta parameters), \code{$best_args} (best meta parameters), and \code{$args} (all meta parameters with cv err)
}
\description{
This function cross validates a grid of meta-parameter arguments implied by \code{expand.grid(...)}. 
The final model is returned with the best fitting parameters.
}
\details{
Non-grid arguments (such as \code{distribution}, \code{weights}, etc) can be passed as normal.
}
