% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_influence.R
\name{importance}
\alias{importance}
\alias{importance.lmerboost}
\alias{importance.mvtb}
\alias{importance.pcb}
\alias{importance.twostage}
\title{Compute variable importance (influence) scores from \code{mvtboost} models.}
\usage{
importance(model, n.trees = NULL, relative = TRUE, sort = TRUE, ...)

\method{importance}{mvtb}(model, n.trees = NULL, relative = "col",
  sort = FALSE, ...)

\method{importance}{twostage}(model, n.trees = NULL, relative = TRUE,
  sort = FALSE, ...)

\method{importance}{lmerboost}(model, n.trees = NULL, relative = TRUE,
  sort = FALSE, ...)

\method{importance}{pcb}(model, n.trees = NULL, relative = "col",
  sort = FALSE, ...)
}
\arguments{
\item{model}{object from \code{mvtb, lmerboost, pcb, twostage}}

\item{n.trees}{number of trees to use. Defaults to the minimum number of trees by CV, test, or training error for each outcome.}

\item{relative}{How to scale the multivariate importances. If \code{"col"}, each column sums to 100. If \code{"tot"}, the whole matrix sums to 100 (a percent). Otherwise, the raw reductions in SSE are returned.}

\item{sort}{whether or not results should be (reverse) sorted. Defaults to FALSE.}

\item{...}{not used}
}
\description{
Compute variable importance (influence) scores from \code{mvtboost} models.
}
\section{Methods (by class)}{
\itemize{
\item \code{mvtb}: mvtb

\item \code{twostage}: twostage

\item \code{lmerboost}: lmerboost

\item \code{pcb}: pcb
}}

