% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_influence.R
\name{influence.mvtb}
\alias{influence.mvtb}
\title{Compute influence scores from mvtb}
\usage{
\method{influence}{mvtb}(model, n.trees = NULL, relative = "col",
  sort = FALSE, ...)
}
\arguments{
\item{model}{\code{mvtb} output model}

\item{n.trees}{number of trees to use. Defaults to the minimum number of trees by CV, test, or training error for each outcome.}

\item{relative}{How to scale the multivariate influences. If \code{"col"}, each column sums to 100. If \code{"tot"}, the whole matrix sums to 100 (a percent). Otherwise, the raw reductions in SSE are returned.}

\item{sort}{whether or not results should be (reverse) sortd. Defaults to FALSE.}

\item{...}{Additional arguments passed to \code{gbm::relative_influence}}
}
\description{
Compute influence scores from mvtb
}

