% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerboost.R
\name{lmerboost}
\alias{lmerboost}
\alias{lmerboost.fit}
\title{Boosted decision trees with random effects}
\usage{
lmerboost(y, X, id, train.fraction = NULL, subset = NULL,
  bag.fraction = 0.5, cv.folds = 1, indep = TRUE, M = 100,
  lambda = 0.01, nt = 1, depth = 5, save.mods = FALSE,
  stop.threshold = 0.001, mc.cores = 1, verbose = TRUE, ...)

lmerboost.fit(y, X, id, train.fraction = NULL, subset = NULL,
  indep = TRUE, M = 100, lambda = 0.01, nt = 1, depth = 5,
  bag.fraction = 0.5, calc.derivs = FALSE, stop.threshold = 0.001,
  verbose = TRUE, save.mods = FALSE, ...)
}
\arguments{
\item{y}{outcome vector (continuous)}

\item{X}{matrix or data frame of predictors}

\item{id}{name or index of grouping variable}

\item{train.fraction}{of sample used for training}

\item{subset}{index of observations to use for training}

\item{bag.fraction}{fraction of training set used at each iteration}

\item{cv.folds}{number of cross-validation folds}

\item{indep}{whether randome effects are independent (default TRUE)}

\item{M}{number of trees}

\item{lambda}{step size}

\item{nt}{number of trees fit at each iteration}

\item{depth}{depth of trees}

\item{save.mods}{whether to save the lmer model at each iteration (required to use 'predict' later)}

\item{stop.threshold}{unused}

\item{mc.cores}{number of parallel cores}

\item{verbose}{whether fitting is verbose}

\item{...}{arguments passed to gbm}

\item{calc.derivs}{whether to calculate derivatives at each iteration with lmer}
}
\description{
Boosted decision trees with random effects
}
\section{Functions}{
\itemize{
\item \code{lmerboost.fit}: Fitting function for lmerboost
}}

