% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_plots.R
\name{mvtb.plot1}
\alias{mvtb.plot1}
\title{Plots the model implied effect of 1 predictor for one outcome.}
\usage{
mvtb.plot1(out, response.no, predictor.no,
  n.trees = min(unlist(out$best.trees)), X = NULL, xlab = NULL, ...)
}
\arguments{
\item{out}{mvtb output object}

\item{response.no}{index of the response variable}

\item{predictor.no}{index of the predictor variable}

\item{n.trees}{desired number of trees (default: best trees)}

\item{X}{predictor. If included, a rug is included with the plot showing the density of the variable.}

\item{xlab}{label of the x variable}

\item{...}{extra arguments are passed to plot. See ?par}
}
\value{
Function is called for it's side effect, a plot.
}
\description{
Plots the model implied effect of 1 predictor for one outcome.
}
\seealso{
plot.gbm, mvtb.perspec, heat.covex
}

