% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtboost.R
\name{mvtb}
\alias{mvtb}
\title{Fitting a Multivariate Tree Boosting Model}
\usage{
mvtb(X = X, Y = Y, n.trees = 100, shrinkage = 0.01,
  interaction.depth = 1, trainfrac = 1, samp.iter = FALSE, bag.frac = 1,
  cv.folds = 1, s = NULL, seednum = NULL, compress = FALSE,
  save.cv = FALSE, alpha = 0.5, cov.discrep = 1, weight.type = 1, ...)
}
\arguments{
\item{X}{data frame of predictors}

\item{Y}{data frame for outcome variables. For best performance, outcome variables should be scaled.}

\item{n.trees}{maximum number of iterations (trees) to be included in the model}

\item{shrinkage}{amount of shrinkage to be applied in each univariate model. This is the gbm shrinkage. Default is .01. Smaller shrinkage values require more iterations to ensure good fit.}

\item{interaction.depth}{fixed depth of trees to be included in the model. A tree depth of 1 are stumps (main effects only), higher tree depths capture higher order interactions.}

\item{trainfrac}{proportion of the sample used for training the multivariate additive model.}

\item{samp.iter}{T/F. If true, draws a new training sample at each iteration of the multivariate model. Default: FALSE.}

\item{bag.frac}{proportion of the training sample used to fit univariate trees for each response at each iteration. Default: .5}

\item{cv.folds}{number of cross validation folds. Default: 1. For larger values, runs k + 1 models, where the k models are run in parallel.}

\item{s}{vector of indices denoting observations to be used for the training sample}

\item{seednum}{integer < 1000 to ensure that results are reproducible}

\item{compress}{T/F. Compress output results list using bzip2 (approx 10\% of original size). Default FALSE.}

\item{save.cv}{T/F. Save all k cv models.}

\item{alpha}{optional argument to select predictors that explain more variance or covariance in outcomes. Variance reductions are weighted by alpha, and covariance reductions are weighted by 1-alpha.}

\item{cov.discrep}{see Details.}

\item{weight.type}{see Details.}
}
\value{
Fitted model. This is a list containing the following elements:

\itemize{
  \item covex - covariance explained in each pair of outcomes by each predictor, in original order.
  \item weights - relative weights at each iteration
  \item wm - raw values of the loss function at each iteration (optionally, truncated at zero)
  \item maxiter - maximum number of iterations run
  \item best.iters - list of the best best iteratino (min MSE test, min MSE cv, last)
  \item finaltree - list of trees in the final model
  \item trainerr - vector of the MSE in the training set at each iteration
  \item testerr - vector of MSE in test set at each iteration
  \item params - arguments
  \item iter.models - list of q gbm models from the first 5 iterations and last.
  \item bestxs - vector of predictors selected at each iteration
  \item bestys - vector of dependent variables selected at each iteration
  \item resid - n x q matrix of residuals after all iterations
  \item init - colMeans(Y)
  \item s - indices of training sample
  \item n - number of observations
  \item ocv - if save.cv=TRUE, returns the CV models.
  \item uncomp - function to uncompress the results if necessary.
}
}
\description{
Builds on gbm (Ridgeway 2013; Friedman, 2001) to fit a univariate tree model for each outcome, selecting predictors at each iteration that explain covariance between the outcomes. The number of trees included in the model can be chosen by minimizing the multivariate mean squared error using cross validation or a test set.
}
\details{
Note that this is a beta version with details subject to change. Any contributions are welcome.
}
\references{
Miller P.J., Lubke G.H, McArtor D.B., Bergeman C.S. (2015) Finding structure in data: A data mining alternative to multivariate multiple regression. Psychological Methods.
}

