% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{mvtb.ri}
\alias{mvtb.ri}
\title{Computes the relative influence of each predictor for each outcome.}
\usage{
mvtb.ri(object, n.trees = NULL, weighted = F, relative = "col")
}
\arguments{
\item{object}{mvtb output object}

\item{n.trees}{number of trees to use}

\item{weighted}{T/F. Reductions in SSE are weighted according the covariance explained by each predictor. See \code{?mvtb} for details.}

\item{relative}{If 'col', each column sums to 100. If 'tot', the whole matrix sums to 100 (a percent). Otherwise, the raw reductions in SSE are returned.}
}
\value{
Matrix of (relative) influences.
}
\description{
The relative influence of a predictor is the reduction in sums of squares attributable to splits on individual predictors.
It is often expressed as a percent (sums to 100).
}
\details{
(Relative) influences are the usual reductions in SSE due to splitting on each predictor.
'Weighted' influences are the raw influences weighted by the covariance explained in all pairs of outcomes by that predictor.
This allows predictor selection to be informed by the covariance explained.
Different weighting types are possible, see \code{?mvtb}.
}

