% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_sep.R
\name{mvtb.sep}
\alias{mvtb.sep}
\title{Separate tuning of multivariate tree boosting models.}
\usage{
mvtb.sep(Y, X, n.trees = 100, shrinkage = 0.01, interaction.depth = 1,
  distribution = "gaussian", train.fraction = 1, bag.fraction = 1,
  cv.folds = 1, keep.data = FALSE, s = NULL, seednum = NULL,
  compress = FALSE, save.cv = FALSE, iter.details = TRUE,
  verbose = FALSE, mc.cores = 1, ...)
}
\description{
Separate tuning of multivariate tree boosting models.
}

