% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{mvtb.summary}
\alias{mvtb.summary}
\title{Computes a summary of the multivariate tree boosting model}
\usage{
mvtb.summary(out)
}
\arguments{
\item{out}{mvtb output object}
}
\value{
Returns the best number of trees, the univariate relative influence of each predictor for each outcome, and covariance explained in pairs of outcomes by each predictor
}
\description{
Computes a summary of the multivariate tree boosting model
}
\seealso{
mvtb.ri, gbm.ri, cluster.covex
}

