% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_plots.R
\name{perf.lmerboost}
\alias{perf.lmerboost}
\title{plot lmerboost performance}
\usage{
perf.lmerboost(x, threshold = 0, lag = 1, ...)
}
\arguments{
\item{x}{lmerboost object}

\item{threshold}{absolute differences in error less than this threshold is optimal}

\item{lag}{lag of the differences in error across iterations}

\item{...}{arguments passed to plot}
}
\description{
plot lmerboost performance
}

