% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_plots.R
\name{plot.lmerboost}
\alias{plot.lmerboost}
\title{Marginal plots for lmerboost objects}
\usage{
\method{plot}{lmerboost}(x, X, id, i.var, n.trees = min(x$best.trees), ...)
}
\arguments{
\item{x}{lmerboost object}

\item{X}{matrix of predictors}

\item{id}{name or index of grouping variable}

\item{i.var}{index or names of variables to plot over (can include id index)}

\item{n.trees}{nubmer of trees (default min(x$best.trees))}

\item{...}{unused}
}
\description{
The fitted values are plotted against one or two predictors. Note that
that this is not a partial dependence plot.
}

