% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_plots.R
\name{plot.mvtb}
\alias{plot.mvtb}
\title{Plots the model implied effect of 1 predictor for one outcome}
\usage{
\method{plot}{mvtb}(x, predictor.no = 1, response.no = 1, n.trees = NULL,
  X = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{mvtb output object}

\item{predictor.no}{index of the predictor variable}

\item{response.no}{index of the response variable}

\item{n.trees}{desired number of trees (default: best trees)}

\item{X}{predictor. If included, a rug is included with the plot showing the density of the variable.}

\item{xlab}{label of the x axis}

\item{ylab}{label of the y axis}

\item{...}{extra arguments are passed to plot. See ?par}
}
\value{
Function is called for it's side effect, a plot of the model implied effect along with its influence computed from \code{gbm}
}
\description{
Plots the model implied effect of 1 predictor for one outcome
}
\seealso{
\code{plot.gbm}, \code{mvtb.perspec}, for other plots, \code{heat.covex} to plot the covariance explained by predictors in a heatmap
}

