% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerboost.R
\name{predict.lmerboost}
\alias{predict.lmerboost}
\title{Prediction for lmerboost objects}
\usage{
\method{predict}{lmerboost}(object, newdata, id, n.trees = NULL, ...)
}
\arguments{
\item{object}{lmerboost object}

\item{newdata}{data frame of new data}

\item{id}{column name or index referring to id variable}

\item{n.trees}{number of trees}

\item{...}{unused}
}
\description{
Prediction for lmerboost objects
}

