% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metb.R
\name{predict.metb}
\alias{predict.metb}
\title{Prediction for metb objects}
\usage{
\method{predict}{metb}(object, newdata, id, n.trees = NULL, ...)
}
\arguments{
\item{object}{metb object}

\item{newdata}{data frame of new data}

\item{id}{column name or index referring to id variable}

\item{n.trees}{number of trees}

\item{...}{unused}
}
\description{
Prediction for metb objects
}
