% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtboost.R
\name{predict.mvtb}
\alias{predict.mvtb}
\title{Predicted values}
\usage{
\method{predict}{mvtb}(out, n.trees, newdata)
}
\arguments{
\item{out}{mvtb object}

\item{n.trees}{number of trees. If a list, returns predictions in a}

\item{newdata}{new matrix of predictors.}
}
\value{
Returns a matrix of predictions for each outcome.
If n.trees is a vector, returns an array, where the third dimension corresponds to the
predictions at a given number of trees.
}
\description{
Predicted values
}

