% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb.R
\name{predict.pcb}
\alias{predict.pcb}
\title{Predicted values from principal components boosting}
\usage{
\method{predict}{pcb}(object, n.trees = NULL, newdata, ...)
}
\arguments{
\item{object}{\code{mvtb} object}

\item{n.trees}{(vector) of the number of trees for each outcome. Defaults to the minimum number of trees by CV, test, or training error for each outcome.}

\item{newdata}{matrix of predictors.}

\item{...}{not used}
}
\description{
Predicted values from principal components boosting
}
