% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{predict.twostage}
\alias{predict.twostage}
\title{Prediction for twostage objects}
\usage{
\method{predict}{twostage}(object, newdata, id, n.trees = NULL, ...)
}
\arguments{
\item{object}{object of class \code{twostage}}

\item{newdata}{data frame of predictors augmented with grouping variable}

\item{id}{character or index of grouping variable in \code{newdata}}

\item{n.trees}{number of trees to use. If NULL, uses the best number of trees in \code{object}}

\item{...}{unused}
}
\description{
Prediction for twostage objects
}

