% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{print.mvtb}
\alias{print.mvtb}
\title{Simple default printing of the mvtb output object}
\usage{
\method{print}{mvtb}(x, ...)
}
\arguments{
\item{x}{mvtb output object}

\item{...}{unused}
}
\description{
Simple default printing of the mvtb output object
}
