% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_cverr.R
\name{summary.gbm.cverr}
\alias{summary.gbm.cverr}
\title{Summarizing gbm.cverr Results}
\usage{
\method{summary}{gbm.cverr}(object, ...)
}
\arguments{
\item{object}{Output from \code{gbm.cverr}}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Calling
\code{summary(gbm.cverr)} produces a data frame with rows corresponding to 
sets of metaparameters and columns that denote for each row,
\item{min.cv.error}{Minimum cross-validation error resulting from the given
set of metaparameters.}
\item{w.index}{The index of the (optional) list of weight vectors
corresponding to the given set of metaparameters. This will be omitted if
a list of weights was not provided to \code{gbm.cverr} through the input
parameter \code{w}.}
\item{var.monotone.index}{The index of the (optional) list of monotonicity
vectors corresponding to the given set of metaparameters. This will be
omitted if a list of weights was not provided to \code{gbm.cverr} through
the input parameter \code{var.monotone}.}
\item{interaction.depth}{The interaction depth corresponding to the
given set of metaparameters.}
\item{n.minobsinnode}{Minimum number of observations in the terminal
nodes of the trees for the given set of metaparameters.}
\item{shrinkage}{The shrinkage parameter corresponding to the
given set of metaparameters.}
\item{bag.fraction}{The fraction of independent training observations 
randomly selected to propose the next tree corresponding to
the given set of metaparameters.}
\item{n.trees}{The optimum number of trees to utilize given the set of
metaprameters denoted in the row. Note that entries in this column will be
marked with '>=' if the boosting procedure was terminated due to time running
out for this set of metaparameters, determined by the user-specified 
\code{max.time} passed to \code{gbm.cverr}}

Sets of metaparameters (rows) are ordered from best (top row) to worst (last 
row) in terms of the resulting cross-validation error.
}
\description{
\code{summary} method for class \code{gbm.cverr}
}
\author{
Daniel B. McArtor (dmcartor@nd.edu)
}
