% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerboost.R
\name{summary.lmerboost}
\alias{summary.lmerboost}
\title{Reports the relative influence of each predictor in an lmerboost models}
\usage{
\method{summary}{lmerboost}(object, ...)
}
\arguments{
\item{object}{lmerboost object}

\item{...}{arguments passed to \code{influence}, e.g. \code{n.trees},
\code{relative}}
}
\description{
Reports the relative influence of each predictor in an lmerboost models
}

