% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metboost.R
\name{summary.metb}
\alias{summary.metb}
\title{Reports the relative influence of each predictor in an metb models}
\usage{
\method{summary}{metb}(object, ...)
}
\arguments{
\item{object}{metb object}

\item{...}{arguments passed to \code{influence}, e.g. \code{n.trees},
\code{relative}}
}
\description{
Reports the relative influence of each predictor in an metb models
}

