% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{summary.mvtb}
\alias{summary.mvtb}
\title{Computes a summary of the multivariate tree boosting model}
\usage{
\method{summary}{mvtb}(out, print = TRUE,
  n.trees = min(unlist(out$best.trees)))
}
\arguments{
\item{out}{mvtb output object}

\item{print}{result (default is TRUE)}
}
\value{
Returns the best number of trees, the univariate relative influence of each predictor for each outcome, and covariance explained in pairs of outcomes by each predictor
}
\description{
Computes a summary of the multivariate tree boosting model
}
\seealso{
\code{mvtb.ri}, \code{gbm.ri}, \code{cluster.covex}
}

