% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{twostage}
\alias{twostage}
\title{Two stage lmer boost}
\usage{
twostage(y, x, id, ..., cv.folds = 3, mc.cores = 1, subset = NULL)
}
\arguments{
\item{y}{outcome}

\item{x}{predictor}

\item{id}{name or index of grouping variable in x}

\item{...}{arguments passed to gbm. Arguments can be passed as vectors, and 
tuning by cross validation will be carried out over expand.grid(...)}

\item{subset}{subset of observations}
}
\description{
Two stage lmer boost
}
\seealso{
\link{gbm_grid}
}

