% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{uncomp.mvtb}
\alias{uncomp.mvtb}
\title{Uncompress a compressed mvtb output object}
\usage{
uncomp.mvtb(object)
}
\arguments{
\item{object}{an object of class \code{mvtb}}
}
\description{
This function uncompresses a compressed mvtb output object. All elements are uncompressed.
}

