% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{weighted.ri}
\alias{weighted.ri}
\title{If \code{"weighted"=TRUE}, the influence is weighted by the covariance explained in all pairs of outcomes by that predictor. 
This allows predictor selection to be informed by the covariance explained. 
Different weighting types are possible, see \code{?mvtb}.}
\usage{
weighted.ri(object, Y, X)
}
\description{
If \code{"weighted"=TRUE}, the influence is weighted by the covariance explained in all pairs of outcomes by that predictor. 
This allows predictor selection to be informed by the covariance explained. 
Different weighting types are possible, see \code{?mvtb}.
}

