\name{wellbeing}
\docType{data}
\alias{wellbeing}
\title{Psychological well-being}
\description{

This data set contains 984 survey responses on the psychological well being (Ryff & Keyes, 1995) of middle-aged and older adults. Psychological well-being has 6 sub-scales - autonomy, environmental mastery (envmast), personal growth (prsgrwth), positive relationships with others (posrel), purpose in life (prpsnlf), and self acceptance. These sub-scales are columns 21:26.

Many potential predictors of psychological well-being are included:


gender, education (educ, with levels high-school, vocational, some-college, graduate degree, other), income (inc, with binned levels), age (18-91), chronic health problems (chrhlth), somatic health problems (somhlth), self-reported health problems (slfhlth), positive affect (posaff), negative affect (negaff), positive social control (pscntrl), control of internal states (cistot), commitment (comit), control (cntrl), challenge (chlng), ego resilience (egores), positive social support from friends/family (pssfrnd, pssfam), perceived stressful emotions (pssemot), perceived stressful problems (pssprob), loneliness (lonlnes)

}
             \usage{wellbeing}
             \format{A data frame of 984 observations and 26 variables (6 outcomes and 20 predictors)}
             \source{...}
             \references{
               ..
             }
             \keyword{datasets}