% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{cluster.covex}
\alias{cluster.covex}
\title{Computing a clustered covariance explained matrix}
\usage{
cluster.covex(out, clust.method = "ward.D", dist.method = "manhattan")
}
\arguments{
\item{out}{mvtb output}

\item{clust.method}{clustering method for rows and columns. See ?hclust}

\item{dist.method}{method for computing the distance between two lower triangluar covariance matrices. See ?dist for alternatives.}
}
\value{
clustered covariance matrix, with rows and columns.
}
\description{
For each pair of predictors, computes the distance between the correlation matrices of the outcomes explained by those predictors.
}

