% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{influence}
\alias{influence}
\alias{influence.mvtb}
\alias{influence.twostage}
\alias{influence.metb}
\alias{influence.pcb}
\title{Compute variable influence (influence) scores from \code{mvtboost} models.}
\usage{
influence(model, n.trees = NULL, relative = TRUE, sort = TRUE, ...)

\method{influence}{mvtb}(model, n.trees = NULL, relative = "col",
  sort = FALSE, ...)

\method{influence}{twostage}(model, n.trees = NULL, relative = TRUE,
  sort = FALSE, ...)

\method{influence}{metb}(model, n.trees = NULL, relative = TRUE,
  sort = FALSE, ...)

\method{influence}{pcb}(model, n.trees = NULL, relative = "col",
  sort = FALSE, ...)
}
\arguments{
\item{model}{object from \code{mvtb, metb, pcb, twostage}}

\item{n.trees}{number of trees to use. Defaults to the minimum number of trees by CV, test, or training error for each outcome.}

\item{relative}{How to scale the multivariate influences. If \code{"col"}, each column sums to 100. If \code{"tot"}, the whole matrix sums to 100 (a percent). Otherwise, the raw reductions in SSE are returned.}

\item{sort}{whether or not results should be (reverse) sorted. Defaults to FALSE.}

\item{...}{not used}
}
\description{
Compute variable influence (influence) scores from \code{mvtboost} models.
}
\section{Methods (by class)}{
\itemize{
\item \code{mvtb}: mvtb

\item \code{twostage}: twostage

\item \code{metb}: metb

\item \code{pcb}: pcb
}}

