% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb_covex.R
\name{mvtb.covex}
\alias{mvtb.covex}
\title{Return the covariance explained matrix}
\usage{
mvtb.covex(Y, X, object, n.trees = NULL, iter.details = F,
  cov.discrep = 1, alpha = 0.5, weight.type = 1)
}
\arguments{
\item{object}{an object of class \code{mvtb}}
}
\description{
Return the covariance explained matrix
}

