% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{mvtb.covex}
\alias{mvtb.covex}
\title{Return the covariance explained matrix}
\usage{
mvtb.covex(object)
}
\arguments{
\item{object}{an object of class \code{mvtb}}
}
\description{
Return the covariance explained matrix
}

