% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtb.R
\name{predict.mvtb}
\alias{predict.mvtb}
\title{Predicted values}
\usage{
\method{predict}{mvtb}(object, newdata, n.trees = NULL, ...)
}
\arguments{
\item{object}{\code{mvtb} object}

\item{newdata}{matrix of predictors.}

\item{n.trees}{(vector) of the number of trees for each outcome. Defaults to the minimum number of trees by CV, test, or training error for each outcome.}

\item{...}{not used}
}
\value{
Returns a matrix or vector of predictions for all outcomes.
}
\description{
Predicted values
}
