% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtboost.R
\name{predict.mvtb}
\alias{predict.mvtb}
\title{Predicted values}
\usage{
\method{predict}{mvtb}(object, n.trees = NULL, newdata, ...)
}
\arguments{
\item{object}{mvtb object}

\item{n.trees}{number of trees. If a list, returns predictions in an array. Defaults to the minimum best tree estimate.}

\item{newdata}{matrix of predictors.}

\item{...}{unused}
}
\value{
Returns a matrix of predictions for each outcome.
If n.trees is a vector, returns an array, where the third dimension corresponds to the
predictions at a given number of trees.
}
\description{
Predicted values
}

