% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtboost.R
\name{predict.mvtb}
\alias{predict.mvtb}
\title{Predicted values}
\usage{
\method{predict}{mvtb}(object, n.trees = NULL, newdata, drop = TRUE, ...)
}
\arguments{
\item{object}{\code{mvtb} object}

\item{n.trees}{number of trees. If a vector, returns predictions in an array. Defaults to the minimum number of trees by CV, test, or training error}

\item{newdata}{matrix of predictors.}

\item{drop}{\code{TRUE/FALSE} Drop any dimensions of length 1}

\item{...}{unused}
}
\value{
Returns an (array, matrix, vector) of predictions for all outcomes. The third dimension corresponds to the 
predictions at a given number of trees, the second dimension corresponds to the number of outcomes.
}
\description{
Predicted values
}

