% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvtb_summary.R
\name{print.mvtb}
\alias{print.mvtb}
\title{Simple default printing of the mvtb output object}
\usage{
\method{print}{mvtb}(out)
}
\arguments{
\item{out}{mvtb output object}
}
\description{
Simple default printing of the mvtb output object
}

