% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostage.R
\name{twostage}
\alias{twostage}
\title{Two stage lmer boost}
\usage{
twostage(y, x, id, subset = NULL, ...)
}
\arguments{
\item{y}{outcome}

\item{x}{predictor}

\item{id}{grouping variable}

\item{subset}{subset of observations}

\item{...}{arguments passed to gbm}
}
\description{
Two stage lmer boost
}

