\name{wellbeing}
\docType{data}
\alias{wellbeing}
\title{Psychological well-being}
\description{

This data set contains 984 survey responses on the psychological well being (Ryff & Keyes, 1995) of middle-aged and older adults from the Notre Dame Study of Health and Well Being (Bergeman & Deboek, 2014). Psychological well-being has 6 sub-scales - autonomy, environmental mastery (envmast), personal growth (prsgrwth), positive relationships with others (posrel), purpose in life (prpsnlf), and self acceptance. These sub-scales are columns 21:26.

The following predictors of psychological well-being are included:

\itemize{
\item gender
\item age (18-91)
\item education (educ, with levels high-school, vocational, some-college, graduate degree, other)
\item income (inc, with binned levels)
\item chronic health problems (chrhlth)
\item somatic health problems (somhlth) 
\item self-reported health problems (slfhlth)
\item positive affect (posaff)
\item negative affect (negaff)
\item positive social control (pscntrl)
\item control of internal states (cistot)
\item commitment (comit)
\item control (cntrl)
\item challenge (chlng)
\item ego resilience (egores)
\item positive social support from friends/family (pssfrnd, pssfam)
\item perceived stressful emotions (pssemot)
\item perceived stressful problems (pssprob)
\item loneliness (lonlnes)
}



}
             \usage{wellbeing}
             \format{A data frame of 984 observations and 26 variables (6 outcomes and 20 predictors)}
             \source{
                          Bergeman, C. S., & Deboeck, P. R. (2014). Trait stress resistance and dynamic stress dissipation on health and well-being: The reservoir model. Research in Human Development, 11(2), 108-125.
             
             }
             \references{

               Ryff, C. D., & Keyes, C. L. M. (1995). The structure of psychological well-being revisited. Journal of Personality and Social Psychology, 69(4), 719.
             }
             \keyword{datasets}