\name{wellbeing}
\docType{data}
\alias{wellbeing}
\title{Psychological well-being}
\description{

This data set contains 985 survey responses on the psychological well being (Ryff & Keyes, 1995) of middle-aged and older adults. Psychological well-being has 6 sub-scales - autonomy, environmental mastery (envmast), personal growth (prsgrwth), positive relationships with others (posrel), purpose in life (prpsnlf), and self acceptance. These sub-scales are columns 18:23.

Many potential predictors of psychological well-being are included:


gender, education (educ, with levels high-school, vocational, some-college, graduate degree, other), income (inc, with binned levels), age (18-91), chronic health problems (chrhlth), somatic health problems (somhlth), self-reported health problems (slfhlth), positive affect (posaff), depression (depr), positive social control (pscntrl), control of internal states (cistot), commitment (comit), control (cntrl), challenge (chlng), ego resilience (egores), positive social support from friends/family (pssfrnd, pssfam).

}
             \usage{wellbeing}
             \format{A data frame of 985 observations and 23 variables.}
             \source{...}
             \references{
               ..
             }
             \keyword{datasets}